[BUGFIX] Ensure selected localized profiles are retrieved for list views
The `list` and `list and detail` plugins have a flexform setting
to select a ordered list of profiles, which are saved as a list
of default language uid values despite they are localized or not,
which is the usual case and the requirement TYPO3 and Extbase is
build on.

Doing a casual extbase query to find records (models) based on the
requires to disable respecting the sys_language_uid and overlaying
records, based on the site language defined overlay setting along
with the additional enforcement option `FallbackForNonTranslated`.

This change ensures to set

```php
  $query->getQuerySettings()->setRespectSysLanguage(false);
```

when a list of selected profiles is given next to adding the
constraint (`IN()`) to ensure that localized profiles and
also  mixed localized profiles are retrieved in this case.

A couple of prepared functional tests are unskipped and the
commented code enabled covering that this issue is fixed.