[TASK] Mitigate implicitly nullable parameter declarations
Since PHP 8.4 implicitly nullable parameter declartions are
deprecated and emitting `E_DEPRECATED` notices.

The recommended way to mitigate these places depends on the
current place, but the spotted places within this extension
can be mitigated by simply changing directly to a nullable
type declaration and considerable non-breaking in this case.

[1] https://php.watch/versions/8.4/implicitly-marking-parameter-type-nullable-deprecated