[TASK] Introduce basc `Build/Scripts/runTests.sh`
This change introduces `Build/Scripts/runTests.sh` build
upon the latest state from the TYPO3 Core Team with only
a first set of basic tools. Will be extended with other
changes.

Most are stripped out, minor blocks are left commented
and will be adopted in one of the following changes.

Implemented suits:

* composer
* composerUpdate
* lintPhp

GitHub action for both core versions are added in the
same step with basic install and linting checks, which
will be extended when adding additional tools.