[TASK] Add static code analyzer `phpstan`
Static code analyzer is added to the repository
and configured with dual core configuration and
baseline support provided by the command wrapper
`Build/Scripts/runTests.sh`. The PHPStan level 8
is relativly high and currently reports some
errors for v11 where not all could be added to
the baseline and v12 crashes due to hard errors.

Based on that, baseline is kept empty for now to
investigate and fix most of the issues first and
enable ci pipeline when green state with a small
baseline is archieved.

It still make sense to add it so we can start to
investigate these issues.

Used command(s):

```shell
composer require --no-update --dev \
  'phpstan/phpstan':'^1.12.21' \
  'phpstan/phpdoc-parser':'^1.29.0' \
  'phpstan/phpstan-phpunit':'^1.4.0' \
  'bnf/phpstan-psr-container':'^1.0.1' \
  'friendsoftypo3/phpstan-typo3':'^0.9.0'
```