[TASK] Automatically register test fixture extensions
It is a common requirement to add fixture extensions to
functional tests to test against multiple differentiating
setups, which requires that the fixture extension composer
namespaces are added to the root composer.json.

That can be a maintenance burden and a composer plugin has
been created to tackle this effort, which is still in kind
of a development phase but already usable for the subset
required in this mono repository and is added now to act
as further preparation to enable functional testing for all
extensions.

Used command(s):

```shell
mv composer.json composer.json.1 \
&& cat <<< $(jq --indent 4 '.extra."sbuerk/fixture-packages" += ["packages/*/Tests/Functional/Fixtures/Extensions/*"]' composer.json.1) > composer.json \
&& composer config allow-plugins.sbuerk/fixture-packages true \
&& COMPOSER_ROOT_VERSION='1.2.x-dev' composer require \
  --dev "sbuerk/fixture-packages":"^0.0.2 <0.1" \
&& rm -rf composer.json.*
```