[TASK] Do not expose class names to the public
In classic modes sensible files are exposed to the
public, usually guarded by `.htaccess` rules, which
is not the in composer mode. We support both modes
for the extension and need to guard access to the
files.

Following construct is used to guard these cases:

```php
if (!defined('TYPO3')) {
  die(__CLASS__);
}
```

This exposes eventually class names to the public
and is not what TYPO3 recommends to use gurding
these files for class mode installation.

This change replaces these construct with a more
generic message:

```php
if (!defined('TYPO3')) {
  die('Not authorized');
}
```

Granted, the message is literally wrong and no status
code returned matching the message, but that way we
keep the nuts safe and warm.