[TASK] Introduce ddev based local development instances
This change aims to introduce a first ddev based local
development setup in a raw first form, extended on the
journey and enhanced over time. Still, adding this now
allows to at least start these instances and having a
development and playground albeit not provisining them
with a workable dataset (yet!). And not having them to
be destroyed before working on main or pull-requests
are a time safer already. And, it can be used to get
first user experiences which may adjust further plans
early.

Instance composer.json files are clones of the root
`composer.json file`, with adjusted path repository
path configuration, removed `sbuerk/fixture-package`
and additionally `typo3/minimal` with respective
version of TYPO3 for the instance.

For both instances, additional `.ddev` web container
docker compose file is added to mount the `packages`
and `packages-dev` folder within the ddev container
so that a relative path requirement works the same
in container and outside.

Following points are not covered and will be added
with dedecated changes over time:

* Implement a generator dev extension/package,  with a generic
  dataset to generate instance data.
* Automatic setup and initial dataset import using a generator
  on first installation / ddev start up.
* Contributor/README documentation how to work with the ddev
  based test instances versus the root tooling setup chain.

Used command(s):

```shell
PROJECT_FOLDER="$( pwd )" \
&& cd "${PROJECT_FOLDER}/ddev-instances/core-12" \
&& ddev config \
  --docroot="public" \
  --database="mariadb:10.11" \
  --php-version="8.1" \
  --project-type="typo3" \
  --project-name="core12-academics-v2" \
  --timezone="Europe/Berlin" \
  --use-dns-when-possible \
  --webserver-type="apache-fpm" \
  --webimage-extra-packages="locales-all" \
&& cd "${PROJECT_FOLDER}/ddev-instances/core-13" \
&& ddev config \
  --docroot="public" \
  --database="mariadb:10.11" \
  --php-version="8.2" \
  --project-type="typo3" \
  --project-name="core13-academics-v2" \
  --timezone="Europe/Berlin" \
  --use-dns-when-possible \
  --webserver-type="apache-fpm" \
  --webimage-extra-packages="locales-all"
```