[TASK] Use cgl check mode with TYPO3 v13
It is essential to execute the `cgl` check based on
the `php-cs-fixer` in dry-run mode, controlled with
the `-n` flag of `Build/Scripts/runTests.sh` to get
failing error code when cgl fixes should be applied.

This has been missed, and for TYPO3 v12 already added
in recent change. Let's make the same change for TYPO3
v13 to have a double check in place.