[DOCS] Enhance monorepo and extension repo files
This change enhances the mono-repository (`academic-extensions`)
`README.md` file adding general `2.x` installation instruction
in development mode and a generic info about ugrades from `1.x`.

For each academic extensions, the `README.md` is extended with
a note how to install `2.x` in composer mode installations and
a remark on upgrades from `1.x` pointing to the `UPGRADE.md`
file. This file is added empty, if not existing yet.

Note already existing `UPGRADE.md` files does not list all changes
which are breaking or deprecated yet, only some of the recent onces.
Will be enhanced in further changes.