[BUGFIX] Ensure created error responses are really returned
The `EXT:academic_persons_edit` profile controller provides
action to display the profile edit form expecting a nullable
profile entity and handly invalid profiles on own it's own,
using the TYPO3 ErrorController to create error pages.

In the past the `ErrorController`  methods directly throw a
exception to return the full-page error content, which has
changed and only returns a `ResponseInterface` object with
the content.

Simply returning the `ErrorController` response as extbase
action response will take the response body content only as
page content part concetinating it with content of other
content elements and rendering the whole page layout around
it leading to invalid HTML and error-page nested within the
page theme. That is how the TYPO3 Content Object Renderer
stack is working, and internally operating on `string` and
not responses.

To enforce the use of the full-page error content only as
displayed page content the `Response` object needs to be
sneaked up into the stack and cutting normal COR rendering
of, which can be done by throwing a specific TYPO3 Exception
with the response attached, designed specifically for this
use-case. This `PropagateResponseException` bubbles up and
will be catched by a PSR-15 Middleware simply returning it.

This improves the situation for the frontend user audience,
recieving at least some error instead of blank content.

The whole error-handling concept in these cases may be
revised in the future, but **should* at least properly
process the currently implemented error handling flow.

Througout the controller multiple occurances of recurring
`method-call-outer-exception-wrap-empty-response-return`
is replaced throwing a propagated exception with the error
handler response object along with the general not logged
in handling to align this at least slightly. The whole
thing needs to be reworked and things passed arround along
with are more well thought handling strategy for the future.