[BUGFIX] Adopt required contacts handling changes for TYPO3 requirements
All contact records are edited in the default language only, therefore
the loading of the contact records might not be very intuitive. The
data handler enforces a translation of inline records for which "l10n_mode"
is set to "exclude".

To keep this record in sync with the record in the default language, all
relevant contact fields must have set "l10n_mode" to "exclude", too. We
need an additional "page" field to store the relation between the page
and the contact records in the respective language.

Finally, the handling in the extbase repository adds the missing piece
to configure extbase towards proper language overlay handling combined
with constraints based on backend selected record ids, which are always
in default language to mitigate side effects recieving records from the
database using `Extbase ORM`.

* Make records in the `tx_academiccontacts4pages_domain_model_contact`
  table translatable.
* Hide `tx_academiccontacts4pages_domain_model_contact` table in lits
  view as editing of contacts should only be done in the page properties.
* Exclude hidden, role and contract columns from translation.
* Show only records in default language in role selection.
* Ensure correct cascading delete setting for pages and contracts.