[TASK] Add academic partner relations to pages
This adds the funactionality to manage partnerships and roles for academic
partners by adding a new tab to the page settings. This allows users to
add partners in partnerships to pages and assign roles to them.
The partnerships are provided for page templates by a data processor.
For flexible displying partnerships in the content of the page, tow
plugins for list and teaser view are provided.

* Add database structure and TCA definitions for partnerships and roles
* Add models and repositories for partnerships and roles
* Add classes for label generation and select item processing
* Add icons and language definitions