[!!!][TASK] Migrate `academic_jobs` plugins to `CType`
For a long time TYPO3 provided the sub-type feature
soley for the `tt_content` table and the core team
had to maintain a bag of painfull special cases in
the codebase without real benefit.

Right from the start it was possible to register
extbase plugins directly as `CType` instead of
the mostly used `list_type` sub-type registration.

Since TYPO3 v13 the list_type sub-type feature is
deprecated and is removed in v14, making the life
of the core team a bit easier. With the deprecation,
a migration upgrade wizard helper is provided, which
extension authors to add migrations in an easy step.

This extension needs to support TYPO3 v12 and v13
and a custom upgrade wizard is provided to process
the migration for 2.x.x directly on all suitable
TYPO3 versions.

Included in this change:

* `list_type` based plugin registration are converted
  to `CType` registration, including new content element
  wizard configuration.
* Group all plugins into a dedicated `academics` content
  element group.
* Provide a upgrade wizard to migrate existing list_type
  plugins to CType.
* Register icon in `Configuration/Icons.php`.
* Add missing language files for translation keys.
* Added basic tests for upgrade wizard.
* Added `typo3/cms-install` as hard dependency due to the
  added upgrade wizard.

Technically breaking the when upgrading in instances should
be low due to the provided upgrade wizard. Except related
extension dealing, xclassing or impacting concrete api's
need to do proper adoptions and are generally not covered
by the breaking policy.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/13.4/Deprecation-105213-TCASubTypes.html