[BUGIX] Use correct queries in PluginContentWizard for `academic-persons`
Upgrade wizard to migrate `academic-persons` content plugins has
been added with a previous change leaving functional tests out,
due to project related time constraints.

It turned out that this was a really bad idea and upgrade wizard
is not working correctly.

This change builds correct queries within the `PluginContentWizard`
and define plugin names in expected form ( `oldName => newName` ).
Functional test are now added for all extension plugins to ensure
working upgrade wizard state.

Upgrade wizards requires `typo3/cms-install` to be present, which
is no longer a hard requirement since TYPO3 v13 and dependency is
now added to `composer.json` and `ext_emconf.php` to take care of
this fact.

Introduced already for 2.0.0 as breaking change changelog entry
is now added to the `UPGRADE.md` to make user of this extension
aware of the requirement to execute upgrade wizards.