[TASK] Improve templating of academic partners
The HTML templates should be as modular as
reasonably possible and should also look appealing
with only bootstrap included. Furthermore they should
be semantically optimal fot them to be as accessible
and slim as possible. Therefore the templates were
adjusted and split into reasonable partials. As the
JavaScript for the map was a bit outdated and thus not
working at all it needed an update as well.