[TASK] Fix backend appearances of fields
The label for the fe_user record type was not showing
because there were upper-, lowercase mistakes which
led to the label not beeing able to be found.
Furthermore the size configuration on the category
field was causing it to be way to small to pick any
categories from within the backend.