[TASK] Improve templating of academic contacts4pages
The HTML templates should be as modular as
reasonably possible and should also look appealing
with only bootstrap included. Furthermore they should
be semantically optimal for them to be as accessible
and slim as possible. Therefore the templates were
adjusted and the partials were removed, because the
content is already present in academic persons which
this extension is heavily relying on.