[BUGFIX] Use correct TCA group translation key in `academic-partners`
Content plugins of all academic extensions should be grouped within
a general `academics` CType group in the TCA select field and based
on that all extensions provides creating that TCA select item group
for the `tt_content.CType` using translation from the own extension.

For `EXT:academic-partners` this failed because the used translation
key in TCA Override missed a colon (`:`) as separator between trans
unit id and the file path.

This is now fixed by adding the missing colon.