[TASK] Improve templating of academic programs
The HTML templates should be as modular as
reasonably possible and should also look appealing
with only bootstrap included. Furthermore they should
be semantically optimal fot them to be as accessible
and slim as possible. Therefore the templates were
adjusted and split into reasonable partials. Furthermore
the locallang files got cleaned up a bit which means
the labels for anything in the backend are now inside
the _be file to clearly differentiate the files content.