[!!!][TASK] Improve templating of academic partners
The HTML templates should be as modular as
reasonably possible and should also look appealing
with only bootstrap included. Furthermore they should
be semantically optimal fot them to be as accessible
and slim as possible. Therefore the templates were
adjusted and split into reasonable partials. Furthermore
the backend_layout and its labels were adjusted to
match the implementation in programs and projects.