[TASK] Improve templating of academic jobs
* Remove extbase configuration for classes as tables are automatically
  mapped to models if they have the same name.

The HTML templates should be as modular as
reasonably possible and should also look appealing
with only bootstrap included. Furthermore they should
be semantically optimal for them to be as accessible
and slim as possible. Therefore the templates were
adjusted and split into reasonable partials. Furthermore
locallang files got cleaned up which means labels for
the backend are now in the correct locallang_be file.
To be able to set the correct header sizes the data of
the content element needed to be passed to the view
inside of the controller.