[BUGFIX]  Re-add storage pages field
All plugins were switched from list_type to CType. For list_type plugins
the page field was autmatically added to the TCA show items. For CType
plugins it has to be added manually.

Also all settings for subtypes_excludelist and subtypes_addlist can be
removed, as they are not needed anymore.