[TASK] Allow access on protected property
As the "abstract" property of the "Partner" model is
protected and therefore can not be accessed without
an appropriate public `get` function to return the
property value, accessing it in templates is not
possible.

Thus a `get` function for the "abstract" property has been added to allow usage in e.g. the list view.