[BUGFIX] Allow more characters for `pageTitleFormat` placeholders
To allow configurable person detail view page titles, a placeholder
based format option has been introduced for the added and dispatch
of a (internal) PSR14 event for each resolved placeholder, which
should help projects to implement custom placeholder replacements.

It just turned out, that the used regular expression to split the
placeholder was to restrictive and preventing to implement custom
event listener to make localization identifier like placeholders
to resolve to translated values, for example placeholders like:

```
%%LLL:EXT:myext/Resources/Private/Languages/locallang.db:identifier%%
```

That discovery made it clear, that the pattern needs to be envolved
to allow more customization and not blocking project or extension
developers, which boils down to:

* Allow space character ` ` to be part of the placeholder identifier.
* Allow double-point `:` to be part of the placeholder indentifier.
* Allow colon `;` to be part of the placeholder identifier.
* Allow dot `.` to be part of the placeholder identifer.
* Allow hyphen `-` to be part of the placeholder identifier.
* Allow slash `/` to be part of the placeholder identifier.
* Allow backslash `\` to be part of the placeholder identifier.

The dispatched event is used to verify that the above mentioned
characters are allowed and regonized for placeholder identifiers,
baked as functional tests.

* [1] https://github.com/fgtclb/academic-extensions/pull/137