[BUGFIX] Use `SvgIconProvider` for svg category icons
`EXT:cagtegory_types` provides a yaml based configuration file to
register category types along with icons in one place and ensures
to register categorys at all related places, for example the TCA
configuration for `sys_category` (type feature) along with icon
classes. It uses registered icon identifiers, which are registered
by the extension `ServiceProvider` during DI build time.

For registration of icons the `IconRegistry::detectIconProvider()`
method has been used to decide which iconProvider to use looking
for `.svg` in provided icon file.

However, the implementation hands over the icon identifier and
not the icon file and the detection returns always the default
`BitmapIconProvider` even for  `SVG` files.

This change passes now the icon to the detect method to get the
correct `SvgIconProvider` for svg icon files, which allows to
render svg in the frontend directly and not using an `image` tag.