[TASK] Allow TypoScript settings to override empty FlexForm settings
TYPO3 provides the ability to define settings for plugin in TypoScript and
also as FlexForm (`pi_flexform`) settings. Extbase `ConfigurationManager`
merged settings from both sources having FlexForm options with the higher
priority, which means that empty FlexForm settings overrides TypoScript
settings.

In many cases it is more suitable that TypoScript settings should override
FlexForm settings when not set (empty) to allow instance wide configuration
and placing plugins in `tt_content` without setting the values on each of
these places. In the past, Extension authors needed to define these cases
and implement that manually for each plugin/action again and again, which
was a pain in the ass.

Since TYPO3 v12.3 the new option `ignoreFlexFormSettingsIfEmpty` [1] gives
the option to tell TYPO3 to use TypoScript settings over the empty FlexForm
counterpart in instances even if the orignal extension does not implemented
it that way or missed to deal with it, like we missed it. The sideeffect is,
that we can ship this configuration directly for options where it is really
obious that TypoScript needs to overrule empty FlexForm settings.

For example the `Detail PageID` (`detailPid`) setting for all plugins related
to display profile short information is a good candidate to have it set once
in TypoScript as a system wide setting and can leave the setting in flexform
empty and still link to the correct detail page.

Following that, we now use `ignoreFlexFormSettingsIfEmpty` for the following
FlexForm options of `EXT:academic_persons` to make them overrideable within
TypoScript over empty FlexForm values:

* settings.detailPid
* settings.pageTitleFormat
* settings.showFields

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.3/Feature-99976-IntroduceignoreFlexFormSettingsIfEmptyExtbaseConfiguration.html