[BUGFIX] Harden `SyncChangesToTranslations` event listener
The `SyncChangesToTranslations` PSR-14 event listener reacts on
the a event dispatched in FE, BE and CLI(BE) context and still
relies having a valid global frontend request available in the
global variable `$GLOBALS['TYPO3_REQUEST']` to pull resolved
site from it, which is the basic information to handle the sync
of profile transations.

It has been revealed, that projects tends to integrate DataHandler
hooks dispatching the very same event without having a correct FE
context set globaly and in that case a `NullSite` will be returned
from the BE request object as `site` attribute incompatible to the
`?Site` return type declaration of `getSite()` method.

The solution needs to be refactoring the whole calling chain and
transport requirete site/siteLanguage information directly with
the corresponding PSR-14 event, which is a bigger rework with a
lot of signature changes across multiple academic extensions.

To mitigate the heavy and nasty PHP TypeError

```
FGTCLB\AcademicPersonsEdit\EventListener\SyncChangesToTranslations::getSite():\
Return value must be of type ?TYPO3\CMS\Core\Site\Entity\Site, \
TYPO3\CMS\Core\Site\Entity\NullSite returned
```

in case the above mentioned missuse hits, this change checks for
NullSite and handes it in the way that no site has been provided.

Overall handling in the EventListener is worked to be more type
and error reluctant as a first and quick bugfix.