[BUGFIX] Streamline extbase model properties and getters
Extbase has quite some requirements related to model properties
and setter/getter methods, special with newer PHP versions and
native types.

That also boils down to the two ways models may be created:

* using new (in manual code) and thus this constructor
* hydrating from the database not calling the constructor

Extbase in general maps data directly to the properies for all
data retrieved from the database, which means for relations
that it will be left out and requires that relations are declared
as nullabe and getter/setter can deal with it.

All of these things has not been respected and are now streamlined
to avoid issues in different flavours and places.