[WIP][TASK] Streamline `EXT:academic_persons_edit` template
It has been considered that the shipped templates for the
academic extensions should be based on bootstrap regarding
used markup to provide a easy and streamlined frontend
styling special in combination with the `EXT:bootstrap`
package.

That makes the development of the extensions easier has
also provides a quick and easy start using the extensions.

To follow that decision, this change streamlines the fluid
templates for `EXT:academic_persons_edit` now.

* Added a bunch of icons for actions of the frontend plugin
  to visualize them.

* Streamlined the fluid template, partial and layout files
  to more logical units to give projects a more granual way
  to adjust parts of the templates. This also uses newly
  added icons for actions.

* The frontend CHKEditor implementation has been updated to
  use a newer version and the custom extension is updated
  in the same step.