[BUGFIX] Change tab addition on pages table
Due to the general tab of pages sometimes having more
than just the title palette, fields or palettes added after the
title palette will always show up in the wrong tab.

This was caused by adding a complete tab after the title
palette, which then causes this split-up.

To adress this issue we now manually add the new tab at the
correct place using plain php functions instead of relying on
the `addToAllTCATypes` function provided by the
`ExtensionManagementUtility`. This had to be done since
this function does not support adding after tabs (yet).