[TASK] Add `publish.yml` workflow file for mono-repository release
After combining all the academic extensions into this
mono-repository it moved forward, which also includes
the publish workflow. For `2.x` creating monorepository
GitHub release next to split repository releleases has
been introduced but never backported to `1.x`.

This is now done with this change, introducing the
mono-repository GitHub release creation when a tag
is pushed.