[BUGFIX] Remove `declare(strict_types=1);` in `ext_emconf.php`
The TYPO3 Extension Repository fails when a extension `ext_emconf.php`
contains the PHP `declare(strict_types=1);` and cannot properly decode
the extension details to handle the upload.

This change removes that statement to make `ext_emconf.php` valid for
an automatic upload to TER using the `typo3/tailor` tool.