[TASK] Only allow a single dateTime format
To use a constant dateTime format the IsoDateTime
class was removed and setting the format was done
by assigning a specific format for the field when
initializing the create action. This way only a single
date format is used, which can be validated more
easily later on. Also this format is the default
format of dates in HTML.