[TASK] Simplify error handling
To simplify the error handling of the job form
validation a general message will be displayed on top
of the form to inform the user of an occurring issue.
Through field appearance those with errors will be
highlighted without further information on the error.
This is the most common simple way. More complex
information on the error and its handling should be
discussed and implemented later on to not block any
upcoming project related tasks.