[TASK] Update image upload
As the `$source` inside the `convertFrom` function of the
`ImageUploadConverter` has been an array and now is an
object of type `TYPO3\CMS\Core\Http\UploadedFile`, the
converter needed to be updated to a compatible TYPO3
v12 version. Therefore it was adjusted to work similar to
the `ProfileImageUploadConverter` of the
`EXT:academic_persons_edit`.