[BUGFIX] Use `SiteFinder` as fallback in `SyncChangesToTranslations`
The PSR-14 EventListener `SyncChangesToTranslations` needs the
related SiteConfiguration for the page and determined it from
the global request object (`$GLOBALS['TYPO3_REQUEST']`), which
has been hardend recently to take `NullSite` as null revealed
in a project implementing a custom DataHandler hook that also
dispatches that event, having a invalid global request in place.

The site configuration object (`Site`) should be part of the
event and passed arround and will be done in a later, dedicated
change.

Looking into the EventListener implementation it stores state,
which is broken in multi-site instances anyway and violates
general best-practice to keep services state-less and PSR-14
event listener are services.

This change modifes the listener implemention now:

* Use `SiteFinder` lookup for the passed profile page id as
  fallback to determine the `Site` object.

* Inlines the `init()` data preparation into the `invoke()`
  method, determine required values directly and passing to
  called methods, making class property state obsolete.

* Replace method calls for class properties with passed
  method arguments, to make the method and properties even
  more obsolete.

* Drop class properties and related internal methods, not
  used anymore with aforementioned changes.

* Harden early skip checks by adding more checks and make
  them visible in single checks and avoid a convulated big
  condition.

Such cleanups may be taken as dedicated changes, but here
it is done to mitigate other issues (bugs) and included
based on that assessment.