[TASK] Render select items through itemsProcFunc
To always be able to ensure that the select items for the
frontend and backend are the same it makes sense to also
maintain them in a single place. Therefore an
itemsProcFunc can be used to render the select items for
the backend and a similar function provides the correct
array for the frontend.