[BUGFIX] Remove incorrect deletion of original file in ImageUploadConverter
Jobs in `EXT:academic_jobs` support attaching a logo image. To handle
file uploads, a custom `ImageUploadConverter` was implemented,
originally derived from `ProfileImageUploadConverter` in
`EXT:academic_persons_edit`, which itself is based on the core
`UploadedFileReferenceConverter` from `EXT:form`[1].

With TYPO3 12, attaching files to Extbase domain models is
streamlined using dedicated API methods[2]. After a file is uploaded
and a `sys_file` is created, a `sys_file_reference` linking the job
record to the file is expected.

Previously, the following method attempted to delete the uploaded
`sys_file` immediately after processing:

```
private function createFileReferenceFromFalFileReferenceObject(CoreFileReference $falFileReference): ExtbaseFileReference
{
  if ($falFileReference->getIdentifier() !== null) {
    try {
        // Delete the current profile image with its file reference.
        $falFileReference->getOriginalFile()->delete();
    } catch (ResourceDoesNotExistException) {
    }
  }
  ...
}
```

This logic is flawed: even when the file and its reference were
correctly created, the condition led to the premature deletion of the
original `sys_file`. As a result, the job logo appeared to upload
successfully, but the file was no longer available. This caused an
error to appear in the backend when editing the job record.

This issue has existed since the initial implementation of the
converter and persisted through several revisions[3].

Rather than attempting to fix the condition, this commit removes the
entire deletion block. Since file creation is directly followed by
reference assignment, the uploaded file remains intact and accessible.

[1] https://github.com/TYPO3/typo3/blob/main/typo3/sysext/form/Classes/Mvc/Property/TypeConverter/UploadedFileReferenceConverter.php
[2] https://docs.typo3.org/m/typo3/reference-coreapi/12.4/en-us/ExtensionArchitecture/Extbase/Reference/FileUpload.html#extbase-fileupload-writing-manual
[3] https://github.com/fgtclb/academic-extensions/commit/7690ddda0f2fdfaf358a5c5b0645d0d18dbff2cd#diff-ddb8c63e750b6aed8b757d0cfc94ad045932fb87f5c454686432a419acced3ba