[TASK] Optimize templating and add new icons
To be able to show the default header fields of TYPO3
elements the `EXT:fluid_styled_content` has been added
as a dependency to reuse the header partials for frontend
header rendering.

New translations have been added to display the labels of
values/properties of the job in the frontend. These are
identical to the ones used in the job creation form and
therefore could be using the same locallang key but this
has not been done yet and could be discussed.

The template logic has been adjusted slightly as some job
properties were not shown in the frontend at all yet. Also
it seem like the contact part was completely missed so it
was added too.

While adding the missing properties to the detail and list
view it was noticed, that not every property had an
assigned and appropriate icon. Therefore a few icons for
these properties have been added to visualize th
differences a bit better.