[TASK] Remove code related to editing an image in the profile
We do not need this specific code as our form is always for creating
a new image and never for editing an existing one.

The `handleNoFileUploaded` function is designed to manage the specific
scenario where a user submits a form for an existing record without
uploading a new file. Its purpose is to preserve the existing
`ExtbaseFileReference` so that the file association isn't lost.

However, there is not even the need for such a check, as the converter
won't be used when no file is attached to the image field.