[TASK] Inject Logger via constructor and remove LoggerAwareTrait
This commit refactors `ImageUploadConverter` to inject the compatible
`LoggerInterface` via the constructor instead of using the
`LoggerAwareTrait`.

Instantiation of the logger with dependency injection is now the
recommended procedure. The `LoggerAwareTrait` and
`LoggerAwareInterface` are discouraged in favor of direct dependency
injection[1].

Additionally, the nullsafe operator (?->) has been removed from the
logger calls, as the logger is now guaranteed to be available after
being injected via the constructor.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/11.4/Feature-95044-SupportAutowiredLoggerInterfaceInjection.html