[TASK] Disable PHP8.5 testing for now
PHP8.5.0beta1 introduced new deprecation, which hits TYPO3,
phpunit and other dependencies for several reasons and the
TYPO3 Core related mitigation have been already been merged
or are in the making.

However, new TYPO3 releases are required to fully have all
dependencies ensured to work with PHP8.5.0beta1.

That is not doable quickly, so disable PHP8.5 testing for
now until a reasonable mitigation can be found to unblock
current work in multiple areas.