[TASK] Introduce class `TcaManipulator` in `EXT:academic_base`
The TCA `general` tab (`div`) may contain more than the title
palette as last palette/item based on the fact that extensions
or sitepackage in projects additional fields or palettes may
be added already.

To add custom tab as first tab after the general tab the
`ExtensionManagementUtility::addToAllTCAtypes()` is used
in extensions, which does not really play well and add a
tab to the after the selected field moving other fields
on that tab into the added tab and leading to an unexpected
configuration.

This change introduces a new `TcaManipulator` class to
`EXT:academic_base` to provide an alternative to modify
TCA in the wished way while keeping modification already
applied by other extension and calls.

Some tests are added to cover the working state of the newly
introduced `TcaManipulator->addToPagesTypeGeneralTab()`.

This change acts as preparation to allow correct TCA manipulation
for multiple academic extensions for pull-request [1]

* [1] https://github.com/fgtclb/academic-extensions/pull/164