[TASK] Centralize file upload handling in EXT:academic_base
This commit introduces a new `FileUploadConverter` in the
`EXT:academic_base` extension. This change centralizes the handling
of file uploads, making the process more modular, consistent and
easier to maintain.

This will reduce the code duplication in `EXT:academic_jobs` and
`EXT:academic_persons_edit`, simplifying the addition of new file
upload features in the future.