[TASK] Introduce centralized `FileUploadConverter` implementation
Since ages projects and extensions needs to implement a custom
extbase TypeConvert to support file upload in extbase related
forms to be attached as file references to extbase models or
DTO's, enforcing to make adoptions for TYPO3 core changes over
and over again.

With TYPO3 v13 [1] a new system has been introduced to provide
that feature directly, which we cannot use due to the support
of TYPO3 v12 and a backport within a extension is not reasonable
based on the required patches, xclasses and similiar deep within
extbase internal API and class constructs.

The academic extensions implemeted already such a custom type
converter and needs it in more places and is the reason why it
is now implemented in `EXT:academic_base` to have a shared and
reusable implementation across all academic extensions requiring
it until TYPO3 v12 support is dropped and the new TYPO3 v13 way
can be used.

`EXT:academic_person_edit` implementation has been used as the
starting point for this implementation and is modernized and
streamlined in the same way to respect the changed TypeConverter
registration with TYPO3 v12 [2] by making the configuration in
`EXT:academic_base/Configuration/Services.yaml`. Until TYPO3 v12
support can be dropped the implementation calls a internal method
to configure the "same" configuration hardcoded  in case 3rd party
code may call the deprecated methods, not doing so for TYPO3 v13
to avoid phpstan issues regarding property type registrations. [3][4]

Implementing this in `EXT:academic_base` now mitigates code
duplication and reduces maintenance or missing to fix the
same thing in multiple places.

Orignally the implementation has been borrowed from `EXT:forms`
and the code base diverged quite a lot meanwhile. Due to time
constraints the improved development of TYPO3 v12/v13 is not
part of this change and will be investigated in follow ups and
is therefore commented as a todo in the class php-docblock.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/13.3/Feature-103511-IntroduceExtbaseFileUploadHandling.html
[2] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Feature-94117-ImproveExtbaseTypeConverterRegistration.html
[3] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Deprecation-94117-RegisterExtbaseTypeConvertersAsServices.html
[4] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Breaking-94117-RegisterExtbaseTypeConvertersAsServices.html

Co-authored-by: Stefan Bürk <stefan@buerk.tech>